Update veterinarian registration

+++ src/Pages/VetClinic/Vet Administrator/NavBarVet.js
@@ -60,17 +60,24 @@ function NavBarVet(props) {
   };
 
   const logoutUser = () => {
-    Axios.delete(`${hostUrl}/logout`, {
-      token: localStorage.getItem("rjwt"),
-    });
 
-    Axios.put(`${hostUrl}/logout/user/vetclinic/${user.vetid}`);
 
-    Axios.post(`${hostUrl}/vetclinic/verified/logout/system/logs`, {
-      name: user.vet_name,
-    });
-    localStorage.clear();
-    window.location.replace("/");
+    Axios.put(`${hostUrl}/logout/user/vetclinic/${user.vetid}`).then(
+      (response) => {
+        if (response.data.message == "Success") {
+          Axios.post(`${hostUrl}/vetclinic/verified/logout/system/logs`, {
+            name: user.vet_name,
+          });
+          Axios.delete(`${hostUrl}/logout`, {
+            token: localStorage.getItem("rjwt"),
+          });
+          localStorage.clear();
+          window.location.replace("/");
+        }
+      }
+    );
+
+
   };
 
   return (
